+++ packages/atomic/cypress/integration/facets/category-facet/category-facet-actions.ts
